requirements.txt
absl-py==1.4.0
aiohttp==3.8.5
aiohttp-cors==0.7.0
aiorwlock==1.3.0
aiosignal==1.3.1
ansiwrap==0.8.4
anyio==3.7.1
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
backoff==2.2.1
backports.functools-lru-cache==1.6.5
beatrix-jupyterlab==2023.814.150030
beautifulsoup4==4.12.2
bleach==6.0.0
blessed==1.20.0
boltons==23.0.0
Brotli==1.0.9
cached-property==1.5.2
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
cloud-tpu-client==0.10
cloudpickle==2.2.1
cmake==3.27.2
colorama==0.4.6
colorful==0.5.5
comm==0.1.4
conda==23.7.3
conda-package-handling==2.2.0
conda_package_streaming==0.9.0
contourpy==1.1.0
cryptography==41.0.3
cycler==0.11.0
Cython==3.0.0
dacite==1.8.1
db-dtypes==1.1.1
debugpy==1.6.7.post1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
distlib==0.3.7
distro==1.8.0
dm-tree==0.1.8
docker==6.1.3
docker-pycreds==0.4.0
docstring-parser==0.15
entrypoints==0.4
exceptiongroup==1.1.3
executing==1.2.0
fastapi==0.101.1
fastjsonschema==2.18.0
filelock==3.12.2
flit_core==3.9.0
fonttools==4.42.1
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.6.0
gcsfs==2023.6.0
gitdb==4.0.10
GitPython==3.1.32
google-api-core==1.34.0
google-api-python-client==1.8.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-cloud-aiplatform==1.31.0
google-cloud-artifact-registry==1.8.3
google-cloud-bigquery==3.11.4
google-cloud-bigquery-storage==2.22.0
google-cloud-core==2.3.3
google-cloud-datastore==1.15.5
google-cloud-language==2.11.0
google-cloud-monitoring==2.15.1
google-cloud-resource-manager==1.10.3
google-cloud-secret-manager==2.17.0
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.60.0
gpustat==1.0.0
greenlet==2.0.2
grpc-google-iam-v1==0.12.6
grpcio==1.57.0
grpcio-status==1.48.2
Gymnasium==0.26.3
gymnasium-notices==0.0.1
h11==0.14.0
htmlmin==0.1.12
httpcore==1.0.2
httplib2==0.22.0
httpx==0.25.2
huggingface-hub==0.17.3
idna==3.4
ImageHash==4.3.1
imageio==2.31.1
importlib-metadata==6.8.0
importlib-resources==6.0.1
ipykernel==6.25.1
ipython==8.14.0
ipython-genutils==0.2.0
ipython-sql==0.5.0
ipywidgets==8.1.0
isoduration==20.11.0
jaraco.classes==3.3.0
jedi==0.19.0
jeepney==0.8.0
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter_client==7.4.9
jupyter_core==5.3.1
jupyter-events==0.7.0
jupyter-http-over-ws==0.0.8
jupyter-lsp==2.2.0
jupyter-server==1.24.0
jupyter-server-mathjax==0.2.6
jupyter_server_proxy==4.0.0
jupyter_server_terminals==0.4.4
jupyterlab==3.4.8
jupyterlab_git==0.42.0
jupyterlab-pygments==0.2.2
jupyterlab_server==2.24.0
jupyterlab-widgets==3.0.8
jupytext==1.15.0
keyring==24.2.0
keyrings.google-artifactregistry-auth==1.1.2
kfp==2.0.1
kfp-pipeline-spec==0.2.2
kfp-server-api==2.0.1
kiwisolver==1.4.4
kubernetes==26.1.0
lazy_loader==0.3
lit==16.0.6
llvmlite==0.40.1
lz4==4.3.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdit-py-plugins==0.4.0
mdurl==0.1.2
mistune==3.0.1
more-itertools==10.1.0
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
multimethod==1.9.1
nb-conda==2.2.1
nb-conda-kernels==2.3.1
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.7.4
nbdime==3.2.0
nbformat==5.9.2
nest-asyncio==1.5.6
networkx==3.1
nltk==3.8.1
notebook==6.5.5
notebook-executor==0.2
notebook_shim==0.2.3
numba==0.57.1
numpy==1.23.5
nvidia-ml-py==11.495.46
oauth2client==4.1.3
oauthlib==3.2.2
openai==1.3.8
opencensus==0.11.2
opencensus-context==0.1.3
opentelemetry-api==1.19.0
opentelemetry-exporter-otlp==1.19.0
opentelemetry-exporter-otlp-proto-common==1.19.0
opentelemetry-exporter-otlp-proto-grpc==1.19.0
opentelemetry-exporter-otlp-proto-http==1.19.0
opentelemetry-proto==1.19.0
opentelemetry-sdk==1.19.0
opentelemetry-semantic-conventions==0.40b0
overrides==7.4.0
packaging==23.1
pandas==2.0.3
pandas-profiling==3.6.6
pandocfilters==1.5.0
papermill==2.4.0
parso==0.8.3
pathtools==0.1.2
patsy==0.5.3
pexpect==4.8.0
phik==0.12.3
pickleshare==0.7.5
Pillow==10.0.0
pip==23.2.1
pkgutil_resolve_name==1.3.10
platformdirs==3.10.0
plotly==5.16.1
pluggy==1.2.0
prettytable==3.8.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
proto-plus==1.22.3
protobuf==3.20.3
psutil==5.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
py-spy==0.3.14
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycosat==0.6.4
pycparser==2.21
pydantic==1.10.12
Pygments==2.16.1
PyJWT==2.8.0
pyOpenSSL==23.2.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0.1
pyzmq==24.0.1
ray==2.6.3
ray-cpp==2.6.3
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.5.2
rouge-score==0.1.2
rpds-py==0.9.2
rsa==4.9
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
safetensors==0.4.0
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.11.2
seaborn==0.12.2
SecretStorage==3.3.3
Send2Trash==1.8.2
sentencepiece==0.1.99
sentry-sdk==1.34.0
setproctitle==1.3.3
setuptools==68.1.2
Shapely==1.8.5.post1
simpervisor==1.0.0
six==1.16.0
smart-open==6.3.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.3.2.post1
SQLAlchemy==2.0.20
sqlparse==0.4.4
stack-data==0.6.2
starlette==0.27.0
statsmodels==0.14.0
sympy==1.12
tabulate==0.9.0
tangled-up-in-unicode==0.2.0
tenacity==8.2.3
tensorboardX==2.6.2.2
terminado==0.17.1
textwrap3==0.9.2
threadpoolctl==3.2.0
tifffile==2023.8.12
tinycss2==1.2.1
tokenizers==0.14.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==2.0.0+cu118
torch-xla==2.0
torchvision==0.15.1+cu118
tornado==6.3.3
tqdm==4.66.1
traitlets==5.9.0
transformers==4.35.0
triton==2.0.0
typeguard==2.13.3
typer==0.9.0
typing_extensions==4.7.1
typing-utils==0.1.0
tzdata==2023.3
Unidecode==1.3.7
uri-template==1.3.0
uritemplate==3.0.1
urllib3==1.26.16
uvicorn==0.23.2
virtualenv==20.21.0
visions==0.7.5
wandb==0.15.12
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.2
wheel==0.41.2
widgetsnbextension==4.0.8
wordcloud==1.9.2
wrapt==1.15.0
yarl==1.9.2
ydata-profiling==4.5.1
zipp==3.16.2
zstandard==0.19.0
